requirements.txt
asgiref==3.7.2
astroid==2.15.5
bcrypt==4.0.1
colorama==0.4.6
contourpy==1.1.0
crispy==0.7.4
cycler==0.11.0
dill==0.3.6
Django==4.2.2
django-crispy-forms==1.12.0
djangorestframework==3.14.0
fabio==2023.6.0
fonttools==4.40.0
h5py==3.9.0
isort==5.12.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
matplotlib==3.7.2
mccabe==0.7.0
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
Pillow==9.5.0
platformdirs==3.7.0
pylint==2.17.4
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
pytz==2023.3
silx==1.1.2
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tzdata==2023.3
wrapt==1.15.0
